libselinux: fall back to a default page size if sysconf() fails
selinux_page_size is used as an allocation size by several callers; a
sysconf(_SC_PAGE_SIZE) failure (-1) or implausible value should not
propagate into those. Clamp to 4096 in that case.

Also change its type to size_t since it is used as such in all except
one place. Convert that usage in validatetrans.c too.

Acked-by: Stephen Smalley <stephen.smalley.work@gmail.com>

+++ libselinux/src/selinux_internal.h
@@ -7,7 +7,7 @@
 #include <stdio.h>
 
 extern int require_seusers;
-extern int selinux_page_size;
+extern size_t selinux_page_size;
 
 /* Make pthread_once optional */
 #pragma weak pthread_once
